bots/alex/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.1
google-generativeai>=0.3.2
twilio>=8.11.0
python-telegram-bot>=20.7
supabase>=2.3.4
websockets>=12.0
pydantic>=2.6.0
requests>=2.31.0
audioop-lts>=0.2.1
